fix destroy

+++ xtd.forms/src/xtd/forms/control.cpp
@@ -34,9 +34,10 @@ const xtd::forms::control xtd::forms::control::null;
 xtd::forms::control::~control() {
   if (this->handle_) {
     native::control_api::unregister_wnd_proc(this->handle_);
-    if (this->instance_.use_count() == 1) native::control_api::del(this->handle_);
-    this->handle_ = 0;
-    this->on_handle_destroyed(xtd::event_args::empty);
+    //if (this->instance_.use_count() == 1) native::control_api::del(this->handle_);
+    //this->handle_ = 0;
+    //this->on_handle_destroyed(xtd::event_args::empty);
+    this->destroy_handle();
   }
 }
 
@@ -128,6 +129,7 @@ void xtd::forms::control::create_handle() {
 
 void xtd::forms::control::destroy_handle() {
   handles_.erase(this->handle_);
+  if (this->instance_.use_count() > 1) return;
   native::control_api::destroy(this->handle_);
   this->handle_ = 0;
   this->on_handle_destroyed(xtd::event_args::empty);
